% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as-methods.R
\name{as}
\alias{as}
\title{As("Experiment", "MSnSet")}
\arguments{
\item{from}{\link{Experiment-class} object}
}
\description{
As("Experiment", "MSnSet")
}
\seealso{
Other Experiment.class: \code{\link{Experiment-class}},
  \code{\link{image,Experiment-method}},
  \code{\link{show,Experiment-method}},
  \code{\link{simulate,Experiment-method}}
}

