% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R, R/image-methods.R, R/show-methods.R, R/simulate-methods.R
\docType{class}
\name{Experiment-class}
\alias{Experiment-class}
\alias{image,Experiment-method}
\alias{show,Experiment-method}
\alias{simulate,Experiment-method}
\title{Experiment class.}
\usage{
\S4method{image}{Experiment}(x, ...)

\S4method{show}{Experiment}(object)

\S4method{simulate}{Experiment}(object, seed = NULL)
}
\arguments{
\item{x}{\link{Experiment-class} object}

\item{...}{currently ignored}

\item{object}{\link{Experiment-class} object}

\item{seed}{an integer for \code{set.seed}}
}
\description{
Details about Experiment class.
}
\details{
The central class of simMSnSet package
}
\section{Methods (by generic)}{
\itemize{
\item \code{image}: Showing simulated experiment as an image

\item \code{show}: Show method for an Experiment object

\item \code{simulate}: Simulates the entire experiment
}}
\section{Slots}{

\describe{
\item{\code{phenotypes}}{Phenotypes and Samples}

\item{\code{proteins}}{Peptides and Proteins}

\item{\code{simFinal}}{simulated final intensities}

\item{\code{simNoise}}{simulated noise of intensity measurements}

\item{\code{simMeans}}{simulated actual fold of changes for each peptide/sample}

\item{\code{simSampleBiases}}{simulated systematic biases due to pipetting errors}

\item{\code{simOutliers}}{indexes of simulated outlying samples}

\item{\code{simMissing}}{index of missing measurements}

\item{\code{simulated}}{A logical flag indicating if the experiment has been simulated or not}
}}
\examples{
# 10 proteins measured across 6 samples (3 controls + 3 cases)
pr <- Proteins(nProt=10)
ph <- Phenotypes(originalPhenotypes=c('Control', 'Control', 'Control',
                            'Case', 'Case', 'Case'))
ex <- Experiment(pr, ph)
ex <- simulate(ex)
image(ex)
}
\seealso{
Other Experiment.class: \code{\link{as}}
}

