% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/experiment.R
\name{Experiment-constructor}
\alias{Experiment}
\alias{Experiment-constructor}
\title{Constructor for Experiment class object}
\usage{
Experiment(proteins, phenotypes)
}
\arguments{
\item{proteins}{an instance of \link{Proteins-class}}

\item{phenotypes}{an instance of \link{Phenotypes-class}}
}
\value{
\link{Experiment-class} object
}
\description{
Constructor for Experiment class object
}

