% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Proteins-class}
\alias{Proteins-class}
\alias{show,Proteins-method}
\alias{simulate,Proteins-method}
\title{Proteins class.}
\usage{
\S4method{simulate}{Proteins}(object, seed = NULL)

\S4method{show}{Proteins}(object)
}
\arguments{
\item{seed}{an integer for \code{set.seed}}
}
\description{
Details about Proteins class.
}
\details{
Class holding the information about the proteins and peptides
}
\section{Methods (by generic)}{
\itemize{
\item \code{simulate}: Simulates the number of peptides per protein
(if not provided) and ionization efficiencies of each peptide.

\item \code{show}: Show method for a Proteins object
}}
\section{Slots}{

\describe{
\item{\code{nProt}}{number of proteins}

\item{\code{nPep}}{number of peptides}

\item{\code{lambda}}{parameter of Poisson distribution controlling the number of
peptides per protein during the simulation}

\item{\code{meanIoinizationIntensity}}{a numeric value of mean peptide intensity}

\item{\code{pepCounts}}{integer vector with number of peptides per each protein}

\item{\code{protNames}}{character vector with generated protein names}

\item{\code{pepNames}}{character vector with generacted peptide names}

\item{\code{ionizationEff}}{numeric vector with simulated ionization efficiencies
for the peptides}

\item{\code{features}}{data.frame containing peptide and protein names}
}}
\examples{
pr <- Proteins(nProt=10)
pr <- simulate(pr)
show(pr)
}

