% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/proteins.R
\name{Proteins-constructor}
\alias{Proteins}
\alias{Proteins-constructor}
\title{Constructor for Proteins class object}
\usage{
Proteins(nProt = 20, propChanging = 0.2, fdrPeptide = 0.05,
  fdrMatching = 0.01, lambda = 2, meanIoinizationIntensity = 1e+06,
  missThreshQuantile = 0.05, missThreshSharpness = 10, noise = 1,
  pepCounts = NULL, ...)
}
\arguments{
\item{nProt}{number of proteins}

\item{lambda}{parameter of Poisson distribution
controlling number of peptides per protein}

\item{meanIoinizationIntensity}{numeric value of an average
peptide ionization intensity}

\item{noise}{measurement noise. So far just a single value for all
peptides/proteins.}

\item{pepCounts}{(optional) integer vector with number of peptides
per each protein. If provided then \code{lambda} is ignored.}
}
\value{
\link{Proteins-class} object
}
\description{
Constructor for Proteins class object
}

