% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{simulate}
\alias{simulate}
\title{Simulates components of quantitative LC-MS data}
\usage{
simulate(object, ...)
}
\arguments{
\item{object}{an instance of simulatable classs. It should contain
some parameter and underlying statistical distribution that
generate random samplings.}
}
\value{
The intention is that it returns the instance of the same class,
         but now with filled slots that needs to be simulated.
}
\description{
Simulates components of quantitative LC-MS data
}

