% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Experiment-class}
\alias{Experiment-class}
\alias{image,Experiment-method}
\alias{show,Experiment-method}
\alias{simulate,Experiment-method}
\title{Experiment class.}
\usage{
\S4method{simulate}{Experiment}(object, seed = NULL)

\S4method{show}{Experiment}(object)

\S4method{image}{Experiment}(x, ...)
}
\arguments{
\item{seed}{an integer for \code{set.seed}}
}
\description{
Details about Experiment class.
}
\details{
The central class of simMSnSet package
}
\section{Methods (by generic)}{
\itemize{
\item \code{simulate}: Simulates the entire experiment

\item \code{show}: Show method for an Experiment object

\item \code{image}: Showing simulated experiment as an image
}}
\section{Slots}{

\describe{
\item{\code{phenotypes}}{Phenotypes and Samples}

\item{\code{proteins}}{Peptides and Proteins}

\item{\code{intensities}}{Simulated intensities}

\item{\code{simulated}}{A logical flag indicating if the experiment has been simulated or not}
}}
\examples{
# 10 proteins measured across 6 samples (3 controls + 3 cases)
pr <- Proteins(nProt=10)
ph <- Phenotypes(original=c('Control', 'Control', 'Control',
                            'Case', 'Case', 'Case'))
ex <- Experiment(pr, ph)
ex <- simulate(ex)
image(ex)
}

