% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R, R/show-methods.R, R/simulate-methods.R
\docType{class}
\name{Phenotypes-class}
\alias{Phenotypes-class}
\alias{show,Phenotypes-method}
\alias{simulate,Phenotypes-method}
\title{Phenotypes class.}
\usage{
\S4method{show}{Phenotypes}(object)

\S4method{simulate}{Phenotypes}(object, seed = NULL)
}
\arguments{
\item{object}{\link{Phenotypes-class} object}

\item{seed}{an integer for \code{set.seed}}
}
\description{
Details about Phenotypes class.
}
\details{
Class holding information about the samples
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Show method for a Phenotype object

\item \code{simulate}: Selects which samples are going to be outliers given
the value of the \code{proportionOutlying} slot.
}}
\section{Slots}{

\describe{
\item{\code{n}}{number of samples}

\item{\code{proportionOutlying}}{proportion of outlying or
erroneusly assigned samples.  It is not uncommong that in
large-scale experiments up to ~5\% samples appear as outliers.}

\item{\code{originalPhenotypes}}{character vector with original or intended
phenotype (group) assignment}

\item{\code{simulatedPhenotypes}}{character vector with phenotype assigments
after pretending that some of the samples are actually outliers}

\item{\code{sampleNames}}{character vector with sample names}

\item{\code{meanFoldChange}}{numeric value controlling the applitude of average fold
of change between the conditions. Used as \code{sdlog} in
\code{rlnorm} function.}

\item{\code{pipettingAccuracy}}{a parameter for generating sample systematic biases}

\item{\code{sampleBiases}}{normalization scaling factors}

\item{\code{simulated}}{logical indicated if the object has been simulated}
}}
\examples{
# 10 samples (5 controls + 5 cases)
# Note, the number of outlying samples is somewhat extreme - 50\%.
# This is for demonstration purpose only.
ph <- Phenotypes(original=c('Ctrl', 'Ctrl', 'Ctrl', 'Ctrl', 'Ctrl',
                            'Case', 'Case', 'Case', 'Case', 'Case'),
                 proportionOutlying=0.5)
ph <- simulate(ph, seed=123)
show(ph)
}

