% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Phenotypes-constructor}
\alias{Phenotypes}
\alias{Phenotypes-constructor}
\title{Constructor for Phenotypes class object}
\usage{
Phenotypes(original = rep(LETTERS[1:3], each = 4), proportionOutlying = 0.1)
}
\arguments{
\item{original}{character vector with original
(intended) phenotype assignments}

\item{proportionOutlying}{proportion of outlying samples.
This will be used in simulation as parameter of Bernoulli distribution.}
}
\value{
\link{Phenotypes-class} object
}
\description{
Constructor for Phenotypes class object
}

