% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R, R/show-methods.R, R/simulate-methods.R
\docType{class}
\name{Proteins-class}
\alias{Proteins-class}
\alias{show,Proteins-method}
\alias{simulate,Proteins-method}
\title{Proteins class.}
\usage{
\S4method{show}{Proteins}(object)

\S4method{simulate}{Proteins}(object, seed = NULL)
}
\arguments{
\item{object}{\link{Proteins-class} object}

\item{seed}{an integer for \code{set.seed}}
}
\description{
Details about Proteins class.
}
\details{
Class holding the information about the proteins and peptides
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Show method for a Proteins object

\item \code{simulate}: Simulates the number of peptides per protein
(if not provided) and ionization efficiencies of each peptide.
}}
\section{Slots}{

\describe{
\item{\code{nProt}}{number of proteins}

\item{\code{fdrPeptide}}{FDR of peptide identifications}

\item{\code{fdrMatching}}{FDR of peptide to spectrum or LC-MS feature matching}

\item{\code{propChanging}}{proportion of proteins that are changed between the
different phenotype groups.  At this point it is independent of
phenotype groups.  It shoud, however, depend on phenotype.}

\item{\code{changingProtIdx}}{index indicating which proteins are changing}

\item{\code{nPep}}{number of peptides}

\item{\code{lambda}}{parameter of Poisson distribution controlling the number of
peptides per protein during the simulation}

\item{\code{meanIoinizationIntensity}}{a numeric value of mean peptide intensity}

\item{\code{noise}}{measurement noise. So far one value for all. May be in the
future convert to signal/noise ratio.}

\item{\code{pepCounts}}{integer vector with number of peptides per each protein}

\item{\code{protNames}}{character vector with generated protein names}

\item{\code{pepNames}}{character vector with generacted peptide names}

\item{\code{ionizationEff}}{numeric vector with simulated ionization efficiencies
for the peptides}

\item{\code{missThreshAbsolute}}{soft threshold in absolute intensity.
It has precedence over missThreshQuantile.}

\item{\code{missThreshQuantile}}{quantile to determine soft threshold value.}

\item{\code{missThreshSharpness}}{parameter controlling the sharpness of the
thresholding step function. The suggested interval for the
sharpness values is [2,100]. At 2 it approximates linear
(not necessarily the best setting). At 100 it is almost a
step-function. The suggested values are >5.}

\item{\code{features}}{data.frame containing peptide and protein names}
}}
\examples{
pr <- Proteins(nProt=10)
pr <- simulate(pr)
show(pr)
}

