% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Proteins-constructor}
\alias{Proteins}
\alias{Proteins-constructor}
\title{Constructor for Proteins class object}
\usage{
Proteins(nProt = 20, lambda = 2, meanIoinizationIntensity = 1e+06,
  pepCounts = NULL, ...)
}
\arguments{
\item{nProt}{number of proteins}

\item{lambda}{parameter of Poisson distribution
controlling number of peptides per protein}

\item{meanIoinizationIntensity}{numeric value of an average
peptide ionization intensity}

\item{pepCounts}{(optional) integer vector with number of peptides
per each protein. If provided then \code{lambda} is ignored.}
}
\value{
\link{Proteins-class} object
}
\description{
Constructor for Proteins class object
}

